refactor(data-cite): ES6ify
